jcc sieve as a real tool; g_frameDelta is u32; 3 branch-layout fixes
The jcc-sequence sieve existed only as prose in
docs/patterns/masked-diff-hides-branch-target.md. It is now
`python -m gruntz.audit.jcc_sieve`: it diffs the ordered (offset, mnemonic,
target) branch list of base vs target out of the objdiff object pairs - where
nothing is masked - for every sub-100% function, and classifies each hit
SIGNEDNESS / POLARITY / OTHER / TOPOLOGY. TOPOLOGY is new and is the shape
`--diff` hides hardest: every mnemonic agrees but a branch lands on a different
block (the SetDimensions bug). Each row also carries both sides' ret counts so
the positive-gate lever can be picked off the row.

First run: 168 hits over 1291 sub-100% fns - SIGNEDNESS 2, TOPOLOGY 7, OTHER 27,
POLARITY 134 (98 pure je/jne). The 2 SIGNEDNESS corroborate the historical 6
minus the 4 already fixed, and SetDimensions (now EXACT) correctly does not
appear.

Both SIGNEDNESS hits were real bugs:

* `g_frameDelta` is **u32**, not i32. Every retail comparison is jb/jbe/jae
  (CPlay::Render `cmp eax,0x12 / jbe`, CMulti::PumpA `jb`, UpdateMgrScroll `jb`,
  CRandomAmbientSound::Update `jb`), the /9 divide in CPlay::DrawWorldFrames is
  `mul`+`shr` not `imul`+`sar`, and every (double) cast lowers as a zero-extended
  `fild qword`. It was declared i32 in FIVE headers and papered over with
  `static_cast<u32>` at 16 sites; the four duplicate externs are gone (consumers
  now include the owner's <Rez/FrameClock.h>) and the 16 casts with them.
  CCreditsState::DrawScrollingCredits 85.33 -> 87.83, frame size now matches.
* `CDDSurface::Blit168`'s LUT-build guard is `cmp edi,g_lut16+0x200 / jl` -
  SIGNED - which a u16* pointer walk can never emit. It is a signed counted loop
  strength-reduced onto the store cursor. 66.01 -> 66.50.

Working the POLARITY half turned up two new families, both documented:

* if-body-owns-the-fallthrough.md - cl5 gives the fall-through to the `if` BODY
  and branches over it, so a shared-exit `goto` lays the two tail blocks in the
  wrong order. Nested negated guards fix it: ProbeFootSoft 98.72 -> 99.99,
  ProbeFootBlocked 99.07 -> 99.99, CTileTriggerLogic::Classify 96.03 -> 98.85.
* literal-comparison-form-survives-o2.md - /O2 keeps the operator you typed, so
  `jle` vs `jl` names it. The inlined MFC CArray::SetSize grow clamp is `>=` not
  `>` (5 open-coded copies; AddFromFile 80.93 -> 81.29, AddFromArray 76.90 ->
  77.24), and SubTable's level loop is `> -1` not `>= 0`.

CORRECTNESS FIX in Classify: retail's `cmp [esi+0x38],1 / jne` in the overflow
arm goes to the shared `or eax,-1` block, so that arm returns -1 EITHER WAY and
the guard only gates the Tick(). We returned +1, i.e. reported "still active" on
a duty cycle that had already expired. A prior note asserted the opposite and
cited "the full retail decode"; the branch TARGET disproves it - which is exactly
what --diff masks.

Also: CDDrawChildGroup::PruneOrphans' Lookup miss arm re-zeroed `found`, so cl
compared the slot against itself; retail's `test eax,eax / je` is a `||`
short-circuit.

MAX gate holds (77.7160 -> 77.7218, all per-fn MAX preserved). The 11 current
dips + 9 lost inline accessors are /O2 ripple from the FrameClock include
landing in Play.h / GameMode.h; exact count unchanged at 2972.

+++ src/Gruntz/Demo.cpp
@@ -108,7 +108,7 @@ i32 CDemo::Render() {
             break;
         }
     }
-    if (static_cast<u32>(g_frameDelta) >= static_cast<u32>(m_520)) {
+    if (g_frameDelta >= static_cast<u32>(m_520)) {
         m_520 = 0;
     } else {
         m_520 -= g_frameDelta;
